a5/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-monitoring==0.31.1
google-cloud-storage==1.16.1